+++ include/events/EventType.h
@@ -7,6 +7,12 @@ namespace events {
 enum class EventType : uint8_t {
     SystemStarted,
     SystemReady,
+    WifiConnecting,
+    WifiConnected,
+    WifiDisconnected,
+    WifiConfigRequired,
+    WifiCredentialsSaved,
+    WebServerStarted,
 };
 
 } // namespace events